[DEV-1290] Refactoring + add return mail
