--- src/subdomains/core/staking/services/crypto-staking.service.ts
@@ -1,43 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { CryptoStaking } from '../entities/crypto-staking.entity';
-import { CryptoStakingRepository } from '../repositories/crypto-staking.repository';
-import { Between, Not } from 'typeorm';
-import { PayoutType } from '../entities/staking-reward.entity';
-
-@Injectable()
-export class CryptoStakingService {
-  constructor(private readonly cryptoStakingRepo: CryptoStakingRepository) {}
-
-  // --- USER --- //
-  async getUserInvests(
-    userId: number,
-    dateFrom: Date = new Date(0),
-    dateTo: Date = new Date(),
-  ): Promise<{ deposits: CryptoStaking[]; withdrawals: CryptoStaking[] }> {
-    const cryptoStaking = await this.cryptoStakingRepo.find({
-      where: [
-        { stakingRoute: { user: { id: userId } }, inputDate: Between(dateFrom, dateTo), isReinvest: false },
-        {
-          stakingRoute: { user: { id: userId } },
-          outputDate: Between(dateFrom, dateTo),
-          payoutType: Not(PayoutType.REINVEST),
-        },
-      ],
-      relations: ['cryptoInput', 'stakingRoute', 'stakingRoute.user'],
-      order: { id: 'ASC' },
-    });
-
-    return {
-      deposits: cryptoStaking.filter(
-        (entry) => entry.inputDate >= dateFrom && entry.inputDate <= dateTo && !entry.isReinvest,
-      ),
-      withdrawals: cryptoStaking.filter(
-        (entry) =>
-          entry.outTxId &&
-          entry.outputDate >= dateFrom &&
-          entry.outputDate <= dateTo &&
-          entry.payoutType !== PayoutType.REINVEST,
-      ),
-    };
-  }
-}
